[#200] Fix SecurityParameter validation
To make this work properly, the `getAllowedJsonTypes` method in `JsonTypeChecker` as moved into `ValidatorBase`, where it was made protected so that subclasses could override it. This left `JsonTypeChecker` nearly empty, so its remainder was also moved into `ValidatorBase`.

Overriding `getAllowedJsonTypes` in `SecurityParmeter` so it would expect an `ArrayNode` then fixed this bug.
